+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -38,6 +38,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `OpenDirectory` / `objc2-open-directory`.
   - `ScreenSaver` / `objc2-screen-saver`.
   - `SystemConfiguration` / `objc2-system-configuration`.
+  - `VideoToolbox` / `objc2-video-toolbox`.
 * Added `#[must_use]` attributes where the C headers have them.
 * Added automatic conversion of `Bool` in external function argument and
   return types (similar to what's done on methods).